+++ R/mock.R
@@ -9,7 +9,7 @@
 #' On exit (regular or error), all functions are restored to their previous state.
 #' This is somewhat abusive of R's internals, and is still experimental, so use with care.
 #'
-#' Primitives (such as \code{\link[base]{interactive}}) cannot be mocked, but this can be
+#' Primitives (such as [base::interactive()]) cannot be mocked, but this can be
 #' worked around easily by defining a wrapper function with the same name.
 #'
 #' @param ... named parameters redefine mocked functions, unnamed parameters