+++ R/test-that.R
@@ -7,7 +7,7 @@
 #' Tests are evaluated in their own environments, and should not affect
 #' global state.
 #'
-#' When run from the command line, tests return \code{NULL} if all
+#' When run from the command line, tests return `NULL` if all
 #' expectations are met, otherwise it raises an error.
 #'
 #' @param desc test name.  Names should be kept as brief as possible, as they
@@ -31,7 +31,7 @@ test_that <- function(desc, code) {
   test_code(desc, code, env = parent.frame())
 }
 
-test_code <- function(test, code, env = test_env()) {
+test_code <- function(test, code, env = test_env(), skip_on_empty = TRUE) {
   get_reporter()$start_test(context = get_reporter()$.context, test = test)
   on.exit(get_reporter()$end_test(context = get_reporter()$.context, test = test))
 
@@ -50,9 +50,9 @@ test_code <- function(test, code, env = test_env()) {
   }
 
   frame <- sys.nframe()
-  frame_calls <- function(start_offset, end_offset) {
+  frame_calls <- function(start_offset, end_offset, start_frame = frame) {
     sys_calls <- sys.calls()
-    start_frame <- frame + start_offset
+    start_frame <- start_frame + start_offset
     structure(
       sys_calls[(start_frame):(length(sys_calls) - end_offset - 1)],
       start_frame = start_frame
@@ -67,7 +67,13 @@ test_code <- function(test, code, env = test_env()) {
   expressions_opt <- getOption("expressions")
   expressions_opt_new <- min(expressions_opt + 500L, 500000L)
 
+  # If no handlers are called we skip: BDD (`describe()`) tests are often
+  # nested and the top level might not contain any expectations, so we need
+  # some way to disable
+  handled <- !skip_on_empty
+
   handle_error <- function(e) {
+    handled <<- TRUE
     # First thing: Collect test error
     test_error <<- e
 
@@ -92,6 +98,7 @@ test_code <- function(test, code, env = test_env()) {
     test_error <<- e
   }
   handle_fatal <- function(e) {
+    handled <<- TRUE
     # Error caught in handle_error() has precedence
     if (!is.null(test_error)) {
       e <- test_error
@@ -107,28 +114,42 @@ test_code <- function(test, code, env = test_env()) {
     register_expectation(e)
   }
   handle_expectation <- function(e) {
+    handled <<- TRUE
     e$expectation_calls <- frame_calls(11, 6)
     register_expectation(e)
     invokeRestart("continue_test")
   }
   handle_warning <- function(e) {
+    handled <<- TRUE
     e$expectation_calls <- frame_calls(11, 5)
     register_expectation(e)
     invokeRestart("muffleWarning")
   }
   handle_message <- function(e) {
+    handled <<- TRUE
     invokeRestart("muffleMessage")
   }
   handle_skip <- function(e) {
-    e$expectation_calls <- frame_calls(11, 2)
+    handled <<- TRUE
+
+    if (inherits(e, "skip_empty")) {
+      # Need to generate call as if from test_that
+      e$expectation_calls <- frame_calls(0, 12, frame - 1)
+    } else {
+      e$expectation_calls <- frame_calls(11, 2)
+    }
+
     register_expectation(e)
     signalCondition(e)
   }
 
   test_env <- new.env(parent = env)
   tryCatch(
-    withCallingHandlers(
-      eval(code, test_env),
+    withCallingHandlers({
+      eval(code, test_env)
+      if (!handled)
+        skip_empty()
+      },
       expectation = handle_expectation,
       skip =        handle_skip,
       warning =     handle_warning,
@@ -141,13 +162,14 @@ test_code <- function(test, code, env = test_env()) {
     skip =  function(e) {}
   )
 
+
   invisible(ok)
 }
 
 #' R package to make testing fun!
 #'
 #' Try the example below. Have a look at the references and learn more
-#' from function documentation such as \code{\link{expect_that}}.
+#' from function documentation such as [expect_that()].
 #'
 #' @details Software testing is important, but, in part because
 #' it is frustrating and boring, many of us avoid it.
@@ -156,11 +178,14 @@ test_code <- function(test, code, env = test_env()) {
 #' and integrates with your existing workflow.
 #'
 #' @section Options:
-#' \code{testthat.use_colours}: Should the output be coloured? (Default:
-#' \code{TRUE}).
+#' - `testthat.use_colours`: Should the output be coloured? (Default:
+#' `TRUE`).
 #'
-#' \code{testthat.summary.max_reports}: The maximum number of detailed test
+#' - `testthat.summary.max_reports`: The maximum number of detailed test
 #' reports printed for the summary reporter (default: 15).
+#'
+#' - `testthat.summary.omit_dots`: Omit progress dots in the summary reporter
+#' (default: `FALSE`).
 #'
 #' @docType package
 #' @name testthat