+++ tests/testthat/test-expect-equality.R
@@ -21,6 +21,13 @@ test_that("expect_equivalent ignores attributes for characters", {
   expect_success(expect_equivalent(x1, x2))
 })
 
+test_that("expect_equal and expect_equivalent pass on ... to compare", {
+  x1 <- 1
+  x2 <- x1 + 1e-6
+
+  expect_success(expect_equal(x1, x2, tol = 1e-5))
+  expect_success(expect_equivalent(x1, x2, tol = 1e-5))
+})
 
 test_that("useful message if objects equal but not identical", {
   f <- function() x