+++ tests/testthat/test-teardown-2.R
@@ -1,0 +1,5 @@
+context("teardown-2")
+
+test_that("teardown is completed by previous test", {
+  expect_false(file.exists("teardown.txt"))
+})